Convert device os-update command to oclif
Change-type: patch
