+++ lib/actions/device.js
@@ -112,5 +112,3 @@ Examples:
 			});
 	},
 };
-
-export { osUpdate } from './device_ts';
