+++ lib/app-capitano.js
@@ -54,7 +54,6 @@ capitano.command(actions.auth.whoami);
 
 // ---------- Device Module ----------
 capitano.command(actions.device.init);
-capitano.command(actions.device.osUpdate);
 
 // ---------- OS Module ----------
 capitano.command(actions.os.versions);
